+++ PostApi/src/main/java/com/example/postapi/repository/CommentRepo.java
@@ -8,25 +8,26 @@ import org.springframework.scheduling.annotation.Async;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigInteger;
 import java.util.List;
 
-public interface CommentRepo extends ReactiveCrudRepository<Comment,Long> {
+public interface CommentRepo extends ReactiveCrudRepository<Comment,BigInteger> {
     @Async
     @Query("SELECT comment_id,post_id,user_id,text FROM comment_post INNER JOIN post ON comment_post.post_id = post.id where comment_post.post_id=:id;")
-    Flux<Comment> findAllByPostId(Long id);
+    Flux<Comment> findAllByPostId(BigInteger id);
     @Async
     @Query("INSERT INTO public.comment_post(post_id, user_id, comment_text) VALUES ( :post_id, :user_id, :text);")
-    Mono<Comment> saveComment(Long post_id,Long user_id, String text);
+    Mono<Comment> saveComment(BigInteger post_id,Long user_id, String text);
     @Async
     @Query("DELETE FROM public.comment_post WHERE post_id=:post_id and user_id = :user_id and text = :text; ")
-    Mono<Void> deleteComment(Long post_id,Long user_id, String text);
+    Mono<Void> deleteComment(BigInteger post_id,Long user_id, String text);
 
     @Async
     @Query("SELECT comment_id,post_id,user_id,text FROM comment_post INNER JOIN post ON comment_post.post_id = post.id where post_id in (:posts_id) LIMIT 10;")
-    Flux<Comment> findAllByManyPosts(List<Long> posts_id);
+    Flux<Comment> findAllByManyPosts(List<BigInteger> posts_id);
     @Async
     @Query("SELECT comment_id,post_id,user_id,text FROM comment_post INNER JOIN post ON comment_post.post_id = post.id where post_id in :posts_id LIMIT 10, OFFSET :start ;")
-    Flux<Comment> findAllByManyPosts(List<Long> posts_id, long start);
+    Flux<Comment> findAllByManyPosts(List<BigInteger> posts_id, long start);
 
 
 }
