+++ PostApi/src/main/java/com/example/postapi/service/CommentService.java
@@ -4,11 +4,12 @@ import com.example.postapi.model.Comment;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface CommentService {
-    public Flux<Comment> getComments(Long post);
-    public Mono<Comment> saveComment(Long post_id, Long user_id, String text);
-    public Mono<Void> deleteComment(Long post_id,Long user_id, String text);
-    public Flux<Comment> findAllByManyPosts(List<Long> posts_id);
+    public Flux<Comment> getComments(BigInteger post);
+    public Mono<Comment> saveComment(BigInteger post_id, Long user_id, String text);
+    public Mono<Void> deleteComment(BigInteger post_id,Long user_id, String text);
+    public Flux<Comment> findAllByManyPosts(List<BigInteger> posts_id);
 }
