+++ PostApi/src/main/java/com/example/postapi/service/CommentServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -15,19 +16,19 @@ public class CommentServiceImpl implements CommentService{
     @Autowired
     private CommentRepo commentRepo;
     @Override
-    public Flux<Comment> getComments(Long post){
+    public Flux<Comment> getComments(BigInteger post){
         return commentRepo.findAllByPostId(post);
     }
     @Override
-    public Mono<Comment> saveComment(Long post_id,Long user_id, String text){
+    public Mono<Comment> saveComment(BigInteger post_id,Long user_id, String text){
         return  commentRepo.saveComment(post_id,user_id,text);
     }
     @Override
-    public Mono<Void> deleteComment(Long post_id,Long user_id, String text){
+    public Mono<Void> deleteComment(BigInteger post_id,Long user_id, String text){
         return commentRepo.deleteComment(post_id,user_id,text);
     }
     @Override
-    public Flux<Comment> findAllByManyPosts(List<Long> posts_id){
+    public Flux<Comment> findAllByManyPosts(List<BigInteger> posts_id){
         return commentRepo.findAllByManyPosts(posts_id);
     }
 }
