+++ PostApi/src/main/java/com/example/postapi/service/LikeService.java
@@ -5,10 +5,12 @@ import com.example.postapi.model.Post;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigInteger;
+
 public interface LikeService {
-    public Flux<Long> findAllByPostId(Long postId);
+    public Flux<Long> findAllByPostId(BigInteger postId);
     public Flux<Long> findAllByUserId(Long userId);
-    public Mono<Like> saveLike(Long postId,Long userId);
-    public Mono<Void> deleteLike(Long postId, Long userId);
-    Mono<Like> findByPostIdAndIdUser(Long postId,Long userId);
+    public Mono<Like> saveLike(BigInteger postId,Long userId);
+    public Mono<Void> deleteLike(BigInteger postId, Long userId);
+    Mono<Like> findByPostIdAndIdUser(BigInteger postId,Long userId);
 }
