+++ PostApi/src/main/java/com/example/postapi/service/LikeServiceImpl.java
@@ -7,6 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+
+import java.math.BigInteger;
+
 @Service
 public class LikeServiceImpl implements LikeService{
     @Autowired
@@ -14,17 +17,17 @@ public class LikeServiceImpl implements LikeService{
 
 
 
-    public Mono<Like> findByPostIdAndIdUser(Long postId, Long userId){
+    public Mono<Like> findByPostIdAndIdUser(BigInteger postId, Long userId){
         return likeRepo.findByPostIdAndIdUser(postId,userId);
     }
 
-    public Mono<Like> saveLike(Long postId,Long userId){
+    public Mono<Like> saveLike(BigInteger postId, Long userId){
         return likeRepo.save(postId,userId);
     }
-    public Mono<Void> deleteLike(Long postId, Long userId){
+    public Mono<Void> deleteLike(BigInteger postId, Long userId){
         return likeRepo.deleteByPostIdAndAndIdUser(postId,userId);
     }
-    public Flux<Long> findAllByPostId(Long postId){
+    public Flux<Long> findAllByPostId(BigInteger postId){
         return likeRepo.findAllByPostId(postId);
     }
     public Flux<Long> findAllByUserId(Long userId){
