+++ src/main/java/com/example/MyNetWork/service/ImageService.java
@@ -5,6 +5,7 @@ import com.example.postapi.model.Post;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -12,5 +13,5 @@ public interface ImageService {
     public boolean upload(MultipartFile file, String path, String fileName);
     public String upload(MultipartFile file);
     public void delete(String name);
-    public void delete(List<Post> posts, Long postId);
+    public void delete(List<Post> posts, BigInteger postId);
 }
