+++ src/main/java/com/example/MyNetWork/service/MessageSender.java
@@ -3,6 +3,7 @@ package com.example.MyNetWork.service;
 import com.example.detailsapi.model.Details;
 import com.example.postapi.model.Post;
 
+import java.math.BigInteger;
 import java.util.Set;
 
 public interface MessageSender {
@@ -10,7 +11,7 @@ public interface MessageSender {
     public void send(Set<Long>  SubscriberId, Long id);
     public void send(Long id);
     public void send(Details details);
-    public void  sendAddOrDeleteLike(Long postId, Long userId, String action);
-    public void sendDeletePostRequest(Long id);
+    public void  sendAddOrDeleteLike(BigInteger postId, Long userId, String action);
+    public void sendDeletePostRequest(BigInteger id);
     public void sendGetAuthorPosts(Long id);
 }
